projects/requirements.txt
altair==4.2.0
anyio==3.6.2
asgiref==3.5.0
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
Automat==20.2.0
autopep8==2.0.0
Babel==2.10.3
backcall==0.2.0
bcrypt==4.0.0
beautifulsoup4==4.11.1
blinker==1.5
bs4==0.0.1
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3e   
colorama==0.4.5
commonmark==0.9.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
cssselect==1.1.0
debugpy==1.6.3
decorator==5.1.1
Deprecated==1.2.13
deta==1.1.0
Django==4.0.2
django-ckeditor==6.2.0
django-debug-toolbar==3.2.4
django-js-asset==2.0.0
django-oauth-toolkit==1.7.0
djangorestframework==3.13.1
dnspython==2.2.1
drf-yasg==1.20.0
ecdsa==0.18.0
email-validator==1.3.0
engineering-notation==0.6.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.0
executing==1.0.0
extra-streamlit-components==0.1.56
fastapi==0.85.2
fastapi_permissions==0.2.7
filelock==3.8.0
Flask==2.2.2
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httptools==0.5.0
hydralit-components==1.0.10
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==5.0.0
incremental==21.3.0
inflection==0.5.1
ipykernel==6.15.2
ipython==8.5.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.1
jose==1.0.0
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jwcrypto==1.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
numpy==1.23.3
oauthlib==3.2.0
opencv-python==4.6.0.66
openpyxl==3.0.10
orjson==3.8.1
outcome==1.2.0
packaging==21.3
pandas==1.4.4
parsel==1.6.0
parso==0.8.3
passlib==1.7.4
patsy==0.5.2
pickleshare==0.7.5
Pillow==9.0.1
plotly==5.10.0
plotly-express==0.4.1
priority==2.0.0
prompt-toolkit==3.0.31
Protego==0.2.1
protobuf==3.20.3
psutil==5.9.2
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
pydeck==0.8.0b3
PyDispatcher==2.0.6
Pygments==2.13.0
PyJWT==2.5.0
Pympler==1.0.1
pymssql==2.2.5
pyodbc==4.0.34
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PythonTurtle==0.3.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==304
PyYAML==6.0
pyzmq==23.2.1
queuelib==1.6.2
requests==2.27.1
requests-file==1.5.1
rich==12.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scipy==1.9.1
Scrapy==2.6.2
semver==2.13.0
service-identity==21.1.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
sqlparse==0.4.2
stack-data==0.5.0
starlette==0.20.4
statsmodels==0.13.2
streamlit==1.13.0
streamlit-aggrid==0.3.3
streamlit-authenticator==0.2.1
streamlit-modal==0.0.5
streamlit-option-menu==0.3.2
tenacity==8.1.0
tk==0.1.0
tk-tools==0.16.0
tkcalendar==1.6.1
tldextract==3.3.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
traitlets==5.3.0
trio==0.22.0
Twisted==22.8.0
twisted-iocpsupport==1.0.2
typing_extensions==4.3.0
tzdata==2021.5
tzlocal==4.2
ujson==5.5.0
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.18.3
validators==0.20.0
w3lib==2.0.1
watchdog==2.1.9
watchfiles==0.18.0
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.2.2
wrapt==1.14.0
wsproto==1.2.0
zipp==3.8.1
zope.interface==5.4.0
